api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.6.1
structlog==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
